LIBFCREPO-1189. Started reorganizing truly common code into a plastron.core module.
https://umd-dit.atlassian.net/browse/LIBFCREPO-1189
