+++ plastron-legacy/src/plastron/commands/delete.py
@@ -2,8 +2,9 @@ import logging
 from email.utils import parsedate_to_datetime
 
 from plastron.commands import BaseCommand
-from plastron.exceptions import RESTAPIException
-from plastron.util import get_title_string, ResourceList, parse_predicate_list
+from plastron.core.exceptions import RESTAPIException
+from plastron.rdf import parse_predicate_list, get_title_string
+from plastron.core.util import ResourceList
 
 logger = logging.getLogger(__name__)
