+++ plastron-legacy/src/plastron/commands/ping.py
@@ -1,5 +1,5 @@
 from plastron.commands import BaseCommand
-from plastron.exceptions import FailureException
+from plastron.core.exceptions import FailureException
 
 
 def configure_cli(subparsers):
