+++ plastron-legacy/src/plastron/core/util.py
@@ -4,23 +4,11 @@ import os
 import re
 import shutil
 import sys
-import urllib
-from argparse import ArgumentTypeError
 from datetime import datetime
 from pathlib import Path
 from tempfile import NamedTemporaryFile
-from typing import Tuple
-from urllib.parse import urlsplit
 
-from paramiko import AutoAddPolicy, SSHClient, SSHException
-from paramiko.config import SSH_PORT
-from rdflib import URIRef, Literal
-from rdflib.util import from_n3
-
-from plastron import namespaces
-from plastron.client import Client, TransactionClient
-from plastron.exceptions import FailureException, RESTAPIException
-from plastron.namespaces import dcterms
+from plastron.client import Client, TransactionClient, RESTAPIException
 
 DEFAULT_LOGGING_OPTIONS = {
     'version': 1,
@@ -67,60 +55,6 @@ DEFAULT_LOGGING_OPTIONS = {
 }
 
 logger = logging.getLogger(__name__)
-
-
-def get_title_string(graph, separator='; '):
-    return separator.join([t for t in graph.objects(predicate=dcterms.title)])
-
-
-def parse_predicate_list(string, delimiter=','):
-    if string is None:
-        return None
-    manager = namespaces.get_manager()
-    return [from_n3(p, nsm=manager) for p in string.split(delimiter)]
-
-
-def uri_or_curie(arg: str) -> URIRef:
-    if arg and (arg.startswith('http://') or arg.startswith('https://')):
-        # looks like an absolute HTTP URI
-        return URIRef(arg)
-    try:
-        term = from_n3(arg, nsm=namespaces.get_manager())
-    except KeyError:
-        raise ArgumentTypeError(f'"{arg[:arg.index(":") + 1]}" is not a known prefix')
-    if not isinstance(term, URIRef):
-        raise ArgumentTypeError(f'"{arg}" must be a URI or CURIE')
-    return term
-
-
-def parse_data_property(p: str, o: str) -> Tuple[URIRef, Literal]:
-    return from_n3(p, nsm=namespaces.get_manager()), Literal(from_n3(o))
-
-
-def parse_object_property(p: str, o: str) -> Tuple[URIRef, URIRef]:
-    predicate = from_n3(p, nsm=namespaces.get_manager())
-    obj = uri_or_curie(o)
-    return predicate, obj
-
-
-def get_ssh_client(sftp_uri, **kwargs):
-    if isinstance(sftp_uri, str):
-        sftp_uri = urlsplit(sftp_uri)
-    if not isinstance(sftp_uri, urllib.parse.SplitResult):
-        raise TypeError('Expects a str or a urllib.parse.SplitResult')
-    ssh_client = SSHClient()
-    ssh_client.load_system_host_keys()
-    ssh_client.set_missing_host_key_policy(AutoAddPolicy)
-    try:
-        ssh_client.connect(
-            hostname=sftp_uri.hostname,
-            username=sftp_uri.username,
-            port=sftp_uri.port or SSH_PORT,
-            **kwargs
-        )
-        return ssh_client
-    except SSHException as e:
-        raise FailureException(str(e)) from e
 
 
 def datetimestamp(digits_only=True):
@@ -191,101 +125,6 @@ def strtobool(val):
         raise ValueError("invalid truth value %r" % (val,))
 
 
-class ResourceList:
-    def __init__(self, client: Client, uri_list=None, file=None, completed_file=None):
-        self.client = client
-        self.uri_list = uri_list
-        self.file = file
-        self.use_transaction = True
-        if completed_file is not None:
-            logger.info(f'Reading the completed items log from {completed_file}')
-            # read the log of completed items
-            fieldnames = ['uri', 'title', 'timestamp']
-            try:
-                self.completed = ItemLog(completed_file, fieldnames, 'uri')
-                logger.info(f'Found {len(self.completed)} completed item(s)')
-            except Exception as e:
-                logger.error(f"Non-standard map file specified: {e}")
-                raise FailureException()
-        else:
-            self.completed = None
-        self.completed_buffer = None
-
-    def get_uris(self):
-        if self.file is not None:
-            if self.file == '-':
-                # special filename "-" means STDIN
-                for line in sys.stdin:
-                    yield line
-            else:
-                with open(self.file) as fh:
-                    for line in fh:
-                        yield line.rstrip()
-        else:
-            for uri in self.uri_list:
-                yield uri
-
-    def get_resources(self, client: Client, traverse=None):
-        repo = client.repo
-        for uri in self.get_uris():
-            if not repo.contains(uri):
-                logger.warning(f'Resource {uri} is not contained within the repository {repo.endpoint}')
-                continue
-            for resource, graph in client.recursive_get(uri, traverse=traverse):
-                yield resource, graph
-
-    def process(self, method, use_transaction=True, traverse=None):
-        self.use_transaction = use_transaction
-        if traverse is not None:
-            predicate_list = ', '.join(p.n3() for p in traverse)
-            logger.info(f"{method.__name__} will traverse the following predicates: {predicate_list}")
-
-        if use_transaction:
-            # set up a temporary ItemLog that will be copied to the real item log upon completion of the transaction
-            self.completed_buffer = ItemLog(
-                NamedTemporaryFile().name,
-                ['uri', 'title', 'timestamp'],
-                'uri',
-                header=False
-            )
-            with self.client.transaction(keep_alive=90) as txn_client:  # type: TransactionClient
-                for resource, graph in self.get_resources(client=txn_client, traverse=traverse):
-                    try:
-                        method(resource, graph)
-                    except RESTAPIException as e:
-                        logger.error(f'{method.__name__} failed for {resource}: {e}: {e.response.text}')
-                        # if anything fails while processing of the list of uris, attempt to
-                        # roll back the transaction. Failures here will be caught by the main
-                        # loop's exception handler and should trigger a system exit
-                        try:
-                            txn_client.rollback()
-                            logger.warning('Transaction rolled back.')
-                            return False
-                        except RESTAPIException:
-                            logger.error('Unable to roll back transaction, aborting')
-                            raise FailureException()
-                txn_client.commit()
-                if self.completed and self.completed.filename:
-                    shutil.copyfile(self.completed_buffer.filename, self.completed.filename)
-                return True
-        else:
-            for resource, graph in self.get_resources(client=self.client, traverse=traverse):
-                try:
-                    method(resource, graph)
-                except RESTAPIException as e:
-                    logger.error(f'{method.__name__} failed for {resource}: {e}: {e.response.text}')
-                    logger.warning(f'Continuing {method.__name__} with next item')
-            return True
-
-    def log_completed(self, uri, title, timestamp):
-        if self.completed is not None:
-            row = {'uri': uri, 'title': title, 'timestamp': timestamp}
-            if self.use_transaction:
-                self.completed_buffer.writerow(row)
-            else:
-                self.completed.writerow(row)
-
-
 class ItemLog:
     def __init__(self, filename, fieldnames, keyfield, header=True):
         self.filename = Path(filename)
@@ -353,3 +192,98 @@ class ItemLog:
 
 class ItemLogError(Exception):
     pass
+
+
+class ResourceList:
+    def __init__(self, client: Client, uri_list=None, file=None, completed_file=None):
+        self.client = client
+        self.uri_list = uri_list
+        self.file = file
+        self.use_transaction = True
+        if completed_file is not None:
+            logger.info(f'Reading the completed items log from {completed_file}')
+            # read the log of completed items
+            fieldnames = ['uri', 'title', 'timestamp']
+            try:
+                self.completed = ItemLog(completed_file, fieldnames, 'uri')
+                logger.info(f'Found {len(self.completed)} completed item(s)')
+            except Exception as e:
+                logger.error(f"Non-standard map file specified: {e}")
+                raise
+        else:
+            self.completed = None
+        self.completed_buffer = None
+
+    def get_uris(self):
+        if self.file is not None:
+            if self.file == '-':
+                # special filename "-" means STDIN
+                for line in sys.stdin:
+                    yield line
+            else:
+                with open(self.file) as fh:
+                    for line in fh:
+                        yield line.rstrip()
+        else:
+            for uri in self.uri_list:
+                yield uri
+
+    def get_resources(self, client: Client, traverse=None):
+        repo = client.repo
+        for uri in self.get_uris():
+            if not repo.contains(uri):
+                logger.warning(f'Resource {uri} is not contained within the repository {repo.endpoint}')
+                continue
+            for resource, graph in client.recursive_get(uri, traverse=traverse):
+                yield resource, graph
+
+    def process(self, method, use_transaction=True, traverse=None):
+        self.use_transaction = use_transaction
+        if traverse is not None:
+            predicate_list = ', '.join(p.n3() for p in traverse)
+            logger.info(f"{method.__name__} will traverse the following predicates: {predicate_list}")
+
+        if use_transaction:
+            # set up a temporary ItemLog that will be copied to the real item log upon completion of the transaction
+            self.completed_buffer = ItemLog(
+                NamedTemporaryFile().name,
+                ['uri', 'title', 'timestamp'],
+                'uri',
+                header=False
+            )
+            with self.client.transaction(keep_alive=90) as txn_client:  # type: TransactionClient
+                for resource, graph in self.get_resources(client=txn_client, traverse=traverse):
+                    try:
+                        method(resource, graph)
+                    except RESTAPIException as e:
+                        logger.error(f'{method.__name__} failed for {resource}: {e}: {e.response.text}')
+                        # if anything fails while processing of the list of uris, attempt to
+                        # roll back the transaction. Failures here will be caught by the main
+                        # loop's exception handler and should trigger a system exit
+                        try:
+                            txn_client.rollback()
+                            logger.warning('Transaction rolled back.')
+                            return False
+                        except RESTAPIException:
+                            logger.error('Unable to roll back transaction, aborting')
+                            raise
+                txn_client.commit()
+                if self.completed and self.completed.filename:
+                    shutil.copyfile(self.completed_buffer.filename, self.completed.filename)
+                return True
+        else:
+            for resource, graph in self.get_resources(client=self.client, traverse=traverse):
+                try:
+                    method(resource, graph)
+                except RESTAPIException as e:
+                    logger.error(f'{method.__name__} failed for {resource}: {e}: {e.response.text}')
+                    logger.warning(f'Continuing {method.__name__} with next item')
+            return True
+
+    def log_completed(self, uri, title, timestamp):
+        if self.completed is not None:
+            row = {'uri': uri, 'title': title, 'timestamp': timestamp}
+            if self.use_transaction:
+                self.completed_buffer.writerow(row)
+            else:
+                self.completed.writerow(row)
