+++ plastron-models/pyproject.toml
@@ -3,8 +3,12 @@ name = "plastron-models"
 version = "4.0.0-dev"
 requires-python = ">= 3.8"
 dependencies = [
+    "edtf_validate",
+    "iso639",
+    "lxml",
     "requests",
     "paramiko",
+    "pillow",
     "rdflib",
 ]
